chore: add templated files for packaging the Operator
The final files for packaging the Operator are the package.yaml, CustomResourceDefinition for the snyk-monitor, and a ClusterServiceVersion that carries the metadata for the Operator.
All of these files will be next used to publish a new version of the Operator to OperatorHub.

The files are checked in instead of auto-generated every time to allow for fine control of the permissions and a lot of the Operator details, which the Operator Community Framework specify that they should be manually reviewed and filled in.